Update _app.js
- Add withAuth and withRedux HOC
- Modify _app.js to use HOC instead
